Add files via upload
o	The project involved a meticulous data collection process, including type conversion and ensuring data consistency. Employing blending techniques, dataset relationships were established to enhance overall coherence.

Statistical analysis was conducted, with Tableau Desktop serving as the key tool for proficient data visualization and communication. The analysis specifically delved into restaurant opening trends, revenue patterns, and the generation of comprehensive reports.

The primary goal was to develop a comprehensive report tailored for Zomato stakeholders. This report focused on individual restaurant performance and identified growth areas within the dataset. Collaboration with team members was integral to the project, allowing for the identification of process improvement opportunities. The collective efforts uncovered significant insights, contributing to the development of an interactive dashboard.

In essence, the project seamlessly integrated data collection, statistical analysis, and visualization techniques to provide Zomato stakeholders with a comprehensive understanding of restaurant performance and key growth opportunities. The collaborative approach with the team ensured a thorough exploration of the dataset, leading to the development of an interactive dashboard for enhanced data interpretation. The project involved a meticulous data collection process, including type conversion and ensuring data consistency. Employing blending techniques, dataset relationships were established to enhance overall coherence. Statistical analysis was conducted, with Tableau Desktop serving as the key tool for proficient data visualization and communication. The analysis specifically delved into restaurant opening trends, revenue patterns, and the generation of comprehensive reports. The primary goal was to develop a comprehensive report tailored for Zomato stakeholders. This report focused on individual restaurant performance and identified growth areas within the dataset. Collaboration with team members was integral to the project, allowing for the identification of process improvement opportunities. The collective efforts uncovered significant insights, contributing to the development of an interactive dashboard. In essence, the project seamlessly integrated data collection, statistical analysis, and visualization techniques to provide Zomato stakeholders with a comprehensive understanding of restaurant performance and key growth opportunities. The collaborative approach with the team ensured a thorough exploration of the dataset, leading to the development of an interactive dashboard for enhanced data interpretation.

o	Skills: Excel Pivot · Data Analysis · Tableau · Presentations · Interactive Visualization · Data Visualization